// BehaviorTreeMainProject/frontend/BehaviorTreeFrontend/src/generated/sidebarCategories.ts
// AUTO-GENERATED by scripts/gen-grammar.mjs
// Sources: BehaviorTree.mc4, CRFTypeDef.mc4, ConcreteBT.mc4, PlanningService.mc4, DynamicBTFlowNode.mc4
// Generated at: 2026-01-17T17:47:01.756Z
// DO NOT EDIT MANUALLY

export const GENERATED_SIDEBAR_CATEGORY_DEFINITIONS = [
  {
    "key": "variables",
    "title": "Blackboard",
    "addLabel": "Add Variable"
  },
  {
    "key": "nodes",
    "title": "Behavior Tree Nodes",
    "addLabel": "Add Behavior Node"
  },
  {
    "key": "decorators",
    "title": "Decorator Templates",
    "addLabel": "Add Decorator"
  },
  {
    "key": "services",
    "title": "Service Templates",
    "addLabel": "Add Service"
  },
  {
    "key": "paramTypes",
    "title": "Parameter Types",
    "addLabel": "Add Parameter Type"
  },
  {
    "key": "paramInstances",
    "title": "Parameter Instances",
    "addLabel": "Add Parameter Instance"
  },
  {
    "key": "predTypes",
    "title": "Predicate Types",
    "addLabel": "Add Predicate Type"
  },
  {
    "key": "predInstances",
    "title": "Predicate Instances",
    "addLabel": "Add Predicate Instance"
  },
  {
    "key": "actions",
    "title": "Action Types",
    "addLabel": "Add Action Type"
  }
] as const;

export const GENERATED_DRAGGABLE_NODE_CATEGORIES = [
  "actionInstances"
] as const;